chatbot-backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart

langchain
langchain-community
langchain-core
langchain-groq
langchain-huggingface
langchain-classic

pinecone-client==3.2.2

langchain-google-genai
sentence-transformers

PyPDF

python-dotenv

pydantic
requests
tqdm

loguru

# MongoDB support
pymongo
motor




